refactor: remove multiply namespace
