+++ hw3/lolainta/include/multiply.hpp
@@ -1,9 +1,8 @@
 #pragma once
 #include <cstddef>
 
-namespace Multiply {
+#include "Matrix.hpp"
 class Matrix;
 Matrix multiply_naive(const Matrix&, const Matrix&);
 Matrix multiply_tile(const Matrix&, const Matrix&, size_t);
 Matrix multiply_mkl(const Matrix&, const Matrix&);
-}  // namespace Multiply
